forecast-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
lightgbm==4.1.0
scikit-learn==1.3.2
joblib==1.3.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
